+++ Changelog.md
@@ -12,6 +12,8 @@
 - Removed `DispatchQueue.performOnMainThread()` in favor of `Task` and `MainActor`.
 - Added `ReleaseTrigger` to get a callback when an object gets released.
 - Added `NotificationCenter.ObserverBag` to have a container for notification observer tokens which get automatically removed from the notification center when released.
+- Added `UIApplication.activeWindowScene` and `UIApplication.keyWindow` to retrieve the current window.
+- Added `ObserverCollection` as an observer bag.
 
 ### v3.1.0